Fixed some bugs and started castling
Fixed a bug with pawns moving 2 spaces not checking that the space 2 in
front of them is empty, also fixed a bug with checkmate not working. I
also added a bunch of instance variables to the BoardState class, I'm
implementing castling in possibleMoves() and possibleK() now. Then I
will implement it in makeMove() and then I will write a method
undoMove() or finish MrDumb.